+++ src/main/java/utility/FigureSupplier.java
@@ -1,17 +1,17 @@
 package utility;
 
+import core.basesyntax.Circle;
+import core.basesyntax.IsoscelesTrapezoid;
+import core.basesyntax.Rectangle;
+import core.basesyntax.RightTriangle;
+import core.basesyntax.Square;
 import java.util.Random;
 import model.Color;
 import model.Figure;
-import model.figure.Circle;
-import model.figure.IsoscelesTrapezoid;
-import model.figure.Rectangle;
-import model.figure.RightTriangle;
-import model.figure.Square;
 
 public class FigureSupplier {
     public static final String DEFAULT_COLOR = "white";
-    public static final int DEFAULT_DATA = 10;
+    public static final double DEFAULT_DATA = 10;
     public static final int MIN_RANDOM_VALUE = 1;
     public static final int MAX_RANDOM_VALUE = 100;
     private final ColorSupplier colorSupplier = new ColorSupplier();
@@ -21,7 +21,7 @@ public class FigureSupplier {
     public Figure getRandomFigure() {
         int figurIndex = random.nextInt(colors.length);
         String randomColor = colorSupplier.getRandomColor();
-        int randomSide = generateRandomSideLength();
+        double randomSide = generateRandomSideLength();
         return switch (figurIndex) {
 
             case 0 -> new Circle(randomColor,
@@ -42,7 +42,7 @@ public class FigureSupplier {
 
     }
 
-    public int generateRandomSideLength() {
+    public double generateRandomSideLength() {
         return random.nextInt(MIN_RANDOM_VALUE, MAX_RANDOM_VALUE);
     }
